Separated tests into test/
