+++ lib/profiler.js
@@ -1,6 +1,3 @@
-import testing from 'testing'
-
-// globals
 var profilers = {};
 var enabled = true;
 
@@ -30,30 +27,13 @@ export function measureFrom(before, key, showEvery) {
 	return diffUs;
 }
 
-function getProfiler(key, showEvery)
+export function getProfiler(key, showEvery)
 {
 	if (!profilers[key] || typeof profilers[key] == 'function')
 	{
 		profilers[key] = new Profiler(key, showEvery);
 	}
 	return profilers[key];
-}
-
-/**
- * Test how long profiling takes.
- */
-function testProfilingProfiler(callback)
-{
-	var runs = 10000;
-	for (var i = 0; i < runs; i++)
-	{
-		var now = start();
-		measureFrom(now, 'fake');
-		measureFrom(now, 'profile');
-	}
-	var stats = getStats('profile');
-	testing.assert(stats.meanTimeUs < 5, 'Profiling should take less than 5 µs, took: ' + stats.meanTimeUs, callback);
-	testing.success(callback);
 }
 
 /**
@@ -83,7 +63,7 @@ export function disable() {
 /**
  * Measure some times, show every few requests.
  */
-function Profiler(name, showEvery)
+export function Profiler(name, showEvery)
 {
 	this.name = name;
 	this.showEvery = showEvery;
@@ -127,41 +107,4 @@ Profiler.prototype.toString = function()
 {
 	return 'profiler for ' + this.name;
 };
-
-/**
- * Test the profiler.
- */
-function testProfiler(callback)
-{
-	var runs = 100000;
-	var now = start();
-	var profiler = new Profiler('first');
-	var before;
-	for (var i = 0; i < runs; i++)
-	{
-		before = start();
-		const elapsedUs = measureFrom(before);
-		testing.assert(elapsedUs, 'measureFrom() should return something', callback);
-		profiler.measure(elapsedUs);
-	}
-	var stats = profiler.getStats();
-	testing.assert(stats, 'No profiler stats', callback);
-	testing.assert(stats.timeUs, 'Profiler stats should not be zero', callback);
-	for (i = 0; i < runs; i++)
-	{
-		before = start();
-		const elapsedUs = measureFrom(before, 'second', runs);
-		testing.assert(elapsedUs, 'measureFrom() should not decrease', callback);
-	}
-	var fromStart = measureFrom(now, 'fromStart');
-	testing.assert(fromStart, 'measureFrom() start should not be zero', callback);
-	testing.success(callback);
-}
-
-export function test(callback) {
-	testing.run([
-		testProfilingProfiler,
-		testProfiler,
-	], callback);
-}
